Fix potential SIGFPE from divide by zero in shaders.
The implementation in LLVM_Util we were relying on for safe div and mod
turned out to have the property that they generated safe results (e.g.,
x/0 returns 0, not inf), but still was capable of generating div-by-zero
exception if the app in which OSL is embedded was trapping SIGFPE.

Instead, switch to a truly safe implementation that only does the div
or mod if the denominator is nonzero. Also enhance it by doing a simple
unchecked div or mod any time the denominator is known to be a nonzero
constant (no need to have the conditional in that case).

LLVM_Util add unchecked op_div(), op_mod(), remove safe_div and safe_mod,
which are now unused and of dubious utility.