oslc bug -- getmatrix failed to note that the matrix argument is written. #618
This could lead to a cascade of optimization errors at runtime.
I have no idea why this never showed up as a problem before today.

In addition to fixing it on the oslc side, add a hacky fix to make sure
that old compiled shaders are retroactively fixed as they are loaded at
runtime.