Helper modes for trying to profile noise and pointcloud (#642)
Is your frame expensive, and you have a hunch it might be because you are
using so much noise() or pointcloud lookups? How can you figure that out?

This patch provides some new helper modes for ShadingSystem:

1. When the "profile" attribute is set, it counts the number of noise
calls and reports the total in the stats.

2. When the "no_noise" attribute is set to nonzero, all noise calls
will be replaced with an inexpensive function that just returns 0 (for
signed noise varieties) or 0.5 (for unsigned noises). You can then time
your render in this mode versus with the regular noise calls, and the
different should give you an estimate of how much time was spent doing
noise math.

3. Attribute "no_pointcloud" similarly short-circuits pointcloud lookups.
This can be useful for trying to estimate how much time was spent
related to pointcloud lookups.