Allow optimization of shader groups with known raytype bits. (#668)
A new ShadingSystem::optimize_group() method takes bit fields describing
which ray types are known to be true and known to be false (not being
specified in either means it's indeterminite at optimize time, will only
be known at execution time).

When compiled in this manner, the runtime optimizer will constant-fold
calls to raytype(name) when the name matches those in the "in" and "out"
lists you have supplied.

This allows an app to specialize a shader for particular ray types (such
as a highly optimized version just for shadow rays).

There is also a new group attribute query "raytype_queries", which reveals
the bit pattern of any raytype() calls mde by any of the shaders within
the group.

As an example, the basic logic for specializing shadow shades and non-shadow
shades looks like this:

    // Make the shader group, however you want
    ShaderGroupRef group = shadingsys->ShaderGroupBegin ("groupname", ...);
    ...
    ShaderGroupEnd ();

    // A second reference-counted shader group starts just referring to the
    // same group.
    ShaderGroupRef group_shad = group;

    // Figure out if the shader group contains any raytype("shadow") queries.
    int rayqueries = 0;
    shadingsys->getattribute (group.get(), "raytype_queries", rayqueries);

    std::string serialized;
    int shadowbit = raytypebit(ustring("shadow"));
    if (rayqueries & shadowbit) {
        // The shader group asks (or might ask) for raytype("shadow"). Let's
        // assume it has different behavior for shadow rays and specialize.
        // First, make another group from the same serialized description.
        shadingsys->getattribute (group.get(), "pickle", serialized);
        group_shad = shadingsys->ShaderGroupBegin ("groupname", "surface", serialized);
        ShaderGroupEnd ();
        // Now specialize the shadow-specific group for shadow rays only:
        shadingsys->optimize_group (group_shad.get(),
                                    shadowbit,   /* ray types on: shadow */
                                    ~shadowbit); /* ray types off: everything else */
        // And specialize the original group for anything but shadow rays:
        shadingsys->optimize_group (group_shad.get(),
                                    0,           /* ray types on -- anything */
                                    shadowbit);  /* ray types off -- shadow */
    } else {
        // The shader group definitely does not check for raytype("shadow"), so
        // there's no point in specializing. Just stick to one version of the
        // shader. It will compile automatically in a generic way whenever we
        // first call shadingsys->execute () with that group. We don't need to
        // do anything special.
    }

    ... shading time: ...
    if (it's a shadow ray)
        shadingsys->execute (context, *group_shad, shaderglobals);
    else /* not a shadow ray */
        shadingsys->execute (context, *group, shaderglobals);

The bad news is that it takes a bit more memory (you're making extra
shader groups) and more time to compile (you're optimizing twice for
many shaders). Nonetheless, I'm finding that I come out ahead.

On my production tests, I'm seeing between 2% and 15% speedup -- very
dependent on the scene, as well as the structure of your shaders. YMMV.

The whole scheme assumes that your shaders take significant shortcuts for
shadow rays, based on explicitly using "if (raytype("shadow")) ...".