Hide symbol visibility for the extern "C" osl_* functions LLVM needs. (732)
We never previously mangled or hid the many functions OSL needs to call
by address, which it figures out with dlsym calls, and therefore those
functions needed "C" linkage (no C++ mangling) and external visibility.
But that was the weak point for completely hiding the library symbols
to avoid clashes between versions or multiple OSL copies.

In this patch, we change the visibility of all those symbols to local
(which we also ensure is hidden by default) and since that would make
the dlsym fail, we explicitly make LLVM aware of the mapping from
llvm::Function* to callable address.