Fix memory leak of ASTNode's in osl compiler. (#743)
Chris discovered that oslc was leaking ASTNode allocations, which was
interfering with some of his sanitizer runs.

Function declarations had the raw pointers squirreled away in the symbol
table without ever keeping a ref-counted version, so they were never
freed. So now we keep a vector of the ref-counted pointers for
ASTfunction_declaration in the OSLCompilerImpl, which all get freed when
the compiler destructs.

Quite independently, we were leaking the nodes associated with certain
metadata, they were just dropped on the floor by the add_meta() method,
which walked through what it was handed and used it to adjust things,
but never freed the pointers. This is fixed by changing add_meta to
force the raw pointers to be converted to ref-counted.

I added some code that for DEBUG builds only will track the construction
and destruction of ASTNodes, and upon program termination will print
messages if more nodes were created than destroyed.