Relax the strict typing of ShadingSystem::Parameter() (#750)
SS::Parameter(), which is used to pass parameter instance values,
is very strict about the types matching exactly. I'm loosening it
somewhat -- by allowing the user to pass a float as the
instance value of a triple (color, point, etc.), and to pass one triple
when a different was the declared parameter. This makes it have similar
flexibility to that afforded by ConnectParameters, and I can't think of
a good reason why one should be any stricter than the other.