Add support for native OSL closures in OptiX (#1028)
The current implementation of the OptiX run time defines its own closure structures that replicate that of the native OSL closure structures. This is likely just legacy code as it was easier to use OptiX types like float3 than add support the native OSL types. Removing this duplication and using the native OSL closures improves the maintainability of the code as you will no longer need duplicate logic if the structures ever change.

* Add support for native OSL closures in OptiX

* Ensure Imathx.h is included before OpenImageIo headers in device code

* Add more intelligent memory alignment logic

* Remove unnecessary ustring.h include