oslc can generate dependency files (#1085)
This patch adds support for `-M`, `-MM`, `-MD`, `-MMD`, `-MF` flags,
akin to how they are used in gcc or clang, to generate dependency
files that can be used by make, ninja, etc., to express header
dependencies of osl files.  For example, output might look like:

    $ oslc -MMD test.osl
    Compiled test.osl -> test.oso
    $ cat test.d
    test.oso: test.osl \
      dist/macosx/share/OSL/shaders/color2.h \
      dist/macosx/share/OSL/shaders/stdosl.h

`-MD` and `-MMD` output a dependency file (all headers for `-MD`, just
user headers for `-MMD`), by default to `<inputbasename>.d`, but
overrideable by `-MF<filename>`.

`-M` and `-MM` are versions of `-MD` and `-MMD` but that print the
resulting dependency list to stdout.

We also adjust the printing of "Compiled foo.osl -> foo.oso" so that
it does not happen in the case of `-E`, `-M`, and `-MM` when other
specific output is expected to comprise the whole of stdout.

Also, I *think* that we now no longer use Boost Wave at all on any
platforms, so I put in an `#error` compiler directive to verify. If
nobody complains, I'll soon excise all of the Boost Wave specific
code.