Improve shader performance with always_inline (#1114)
* Make sure that LLVM ops are marked as always_inline

This allows LLVM to truly inline the functions when the optimization level is set to 1 or higher

* Small improvements to llvm_output_bitcode feature

This patches uses underscores instead of dots in the filename to avoid the file from being "hidden" if it started with a "/". Also handle the ":" character which shows up in some material naming conventions and would likely be problematic on windows.

Also add a message to indicate when the bitcode files are written, and make sure the message shows up immediately instead of being deferred.

* Change default LLVM JIT optimization level to 1

The default set of optimization passes produces better code than our previous hand-picked set of passes when combined with the always_inline annotation. Surprisingly, it even reduces the total JIT time as well.

Also fix an alignment issue in llvm_gen_print on the GPU revealed by moving to the more complete set of LLVM optimization passes